Move core/stack to stack/stack
