+++ analyzer/funAnalyzer.go
@@ -7,8 +7,8 @@ import (
 	"zyro/code/mod"
 	"zyro/code/types"
 	"zyro/code/wrapper"
-	"zyro/core/stack"
 	"zyro/logger"
+	"zyro/stack"
 	"zyro/token"
 	"zyro/tokenUtil/splitter"
 	"zyro/util"
