+++ analyzer/methodAnalyzer.go
@@ -4,7 +4,7 @@ import (
 	"zyro/code"
 	"zyro/code/mod"
 	"zyro/code/wrapper"
-	"zyro/core/stack"
+	"zyro/stack"
 	"zyro/token"
 )
