+++ analyzer/staticAnalyzer.go
@@ -7,8 +7,8 @@ import (
 	"zyro/ast"
 	"zyro/code/types"
 	"zyro/code/wrapper"
-	"zyro/core/stack"
 	"zyro/logger"
+	"zyro/stack"
 )
 
 // A regex to match camelCase variable names
