+++ analyzer/typeAnalyzer.go
@@ -5,8 +5,8 @@ import (
 	"zyro/code/mod"
 	"zyro/code/types"
 	"zyro/code/wrapper"
-	"zyro/core/stack"
 	"zyro/logger"
+	"zyro/stack"
 	"zyro/token"
 )
