+++ stack/stack.go
